multiple stores in svelte

+++ QuickQuiz.API/Network/Game/GamePlayer.cs
@@ -7,7 +7,7 @@ namespace QuickQuiz.API.Network.Game
     public class GamePlayer : NetworkPlayer
     {
         public string CategoryVoteId;
-        public float Points;
+        public double Points;
 
         public List<bool> RoundAnswers = new List<bool>(5);
         public List<TimeSpan> AnswerTimes = new List<TimeSpan>(5);
